--- a/pennylane/ops/__init__.py
+++ b/pennylane/ops/__init__.py
@@ -54,8 +54,8 @@ class Identity(CVObservable):
     ev_order = 1
     eigvals = np.array([1, 1])
 
-    @staticmethod
-    def _matrix(*params):
+    @classmethod
+    def _matrix(cls, *params):
         return np.eye(2)
 
     @staticmethod