--- a/pennylane/ops/qubit.py
+++ b/pennylane/ops/qubit.py
@@ -1457,6 +1457,7 @@ ops = {
     "BasisState",
     "QubitStateVector",
     "QubitUnitary",
+    "DiagonalQubitUnitary",
 }