… .github/CHANGELOG.md
… doc/development/plugins.rst
… pennylane/optimize/adagrad.py
… pennylane/optimize/adam.py
… pennylane/optimize/rms_prop.py
… pennylane/plugins/default_gaussian.py
… pennylane/plugins/default_qubit.py
… pennylane/templates/embeddings/amplitude.py
… pennylane/templates/state_preparations/mottonen.py
… tests/gate_data.py
… tests/ops/test_qubit_ops.py
… tests/test_default_qubit.py
… tests/test_operation.py