**Context:**
Default.qubit offers many possibilities to improve the performance. 

**Description of the Change:**
* Change the implementation of matrix-vector products inside default.qubit
* Add `DiagonalGate` for gates that can be applied more easily
* Use tensors to represent state inside `default.qubit`

**Attempted changes that had no large impact:**
* Import matrices at loading of `default.qubit`

**Benefits:**
Speed!

**Possible Drawbacks:**
None

**Related GitHub Issues:**
None

**Todo:**
- [x] Add `DiagonalUnitary` for diagonal gates (e.g. CCZ -> IQP)
- [x] Use `math` instead of `numpy` for scalar operations: 
    ```
    %timeit math.sin(0.3 * math.pi)
    195 ns ± 18.8 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
    %timeit np.sin(0.3 * np.pi)
    672 ns ± 32.6 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
    ```
- [x] Hardcode matrices
```
In [6]: %timeit math.cos(theta / 2) * np.eye(2) + 1j * math.sin(-theta / 2) * xmatrix
5.59 µs ± 167 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [7]: %timeit np.array([[math.cos(theta/2), 1j * math.sin(-theta/2)],[1j * math.sin(-theta/2), math.cos(theta/2)]])
1.96 µs ± 81.1 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [8]: %timeit c = math.cos(theta/2);s=math.sin(theta/2);np.array([[c, 1j * s],[1j * s, c]])
1.72 µs ± 87.4 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [9]: %timeit c = math.cos(theta/2);s=1j * math.sin(theta/2);np.array([[c, s],[s, c]])
1.63 µs ± 121 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
Factor 3 for small matrices
```
In [9]: %timeit block_diag(np.eye(2), _matrix(theta))
34.3 µs ± 572 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [12]: %timeit c = math.cos(theta / 2); js = 1j * math.sin(-theta / 2); np.array([[1, 0, 0, 0], [0,1,0,0],[0,0,c, js], [0,0,js, c]])
3.02 µs ± 79.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
Factor 10 for controlled matrices.
Also Factor 4 for Rot.

**Comparison Checkpoints:**
- 435f482 Use `vec_vec_product` for diagonal gates
- 3a5ef85 Use `einsum` for small gates
- 48750ed Use tensors to represent states inside default.qubit
- 38c4b9e Use inbuilt math for scalar functions (`math`, `cmath`)
- 373c5dc Hardcode matrices
- e9ffb470 Hardcode `math.sqrt(2)`
- 77202d2e Make `matrix` attribute for self-inverse operations

Test with (script is currently on branch `benchmark_commits`, see #568)
```
python benchmark_revisions.py -r master,435f482,3a5ef85,48750ed -d default.qubit -w 5 time bm_entangling_layers
```