@johannesjmeyer 2020-04-15
Here are the benchmarks with the effect of the changes:
## Entangling Layers Benchmark
```
>>> master
Timing per loop: [0.56730501 0.56466833 0.54787934 0.55268263 0.53701176]
>>> 435f482 (Use vec_vec_product for diagonal gates)
Timing per loop: [0.51992123 0.51602465 0.54403876 0.52784254 0.58891966]
>>> 3a5ef85 (Use einsum for 1 and 2 qubit gates)
Timing per loop: [0.45121952 0.46046541 0.41216325 0.41468462 0.43323275]
>>> 48750ed (Use tensors to represent states)
Timing per loop: [0.42049386 0.43040028 0.4063169  0.42034827 0.43499973]
>>> 38c4b9e (Use inbuilt scalar math)
Timing per loop: [0.39011277 0.38066364 0.42326635 0.39533586 0.4167642 ]
>>> 373c5dc (Hardcode matrices)
Timing per loop: [0.31169557 0.30529634 0.27838911 0.27758297 0.2910186 ]
>>> default_qubit_performance (Add DiagonalQubitUnitary, replace np.sqrt)
Timing per loop: [0.28788649 0.27796841 0.28076874 0.27799456 0.29225529]
```

## IQP Circuit Benchmark
```
>>> master
Timing per loop: [0.0289245  0.02738214 0.03033473 0.02664688 0.02840538]
>>> 435f482 (Use vec_vec_product for diagonal gates)
Timing per loop: [0.02706734 0.02178523 0.02321999 0.02706252 0.02449127]
>>> 3a5ef85 (Use einsum for 1 and 2 qubit gates)
Timing per loop: [0.02483001 0.02542964 0.02396653 0.02639955 0.0253584 ]
>>> 48750ed (Use tensors to represent states)
Timing per loop: [0.0277114  0.03146629 0.02625932 0.02402875 0.02867192]
>>> 38c4b9e (Use inbuilt scalar math)
Timing per loop: [0.01664415 0.01638892 0.01558892 0.01653581 0.01715614]
>>> 373c5dc (Hardcode matrices)
Timing per loop: [0.01495848 0.01455966 0.01534762 0.01745174 0.01725933]
>>> default_qubit_performance  (Add DiagonalQubitUnitary, replace np.sqrt)
Timing per loop: [0.01509357 0.01454216 0.0144417  0.01479546 0.01539704]
```
___
You can recreate the benchmarks by running
```
python benchmark_revisions.py -r master,435f482,3a5ef85,48750ed,38c4b9e,373c5dc,default_qubit_performance -w 7 -d default.qubit time bm_entangling_layers
python benchmark_revisions.py -r master,435f482,3a5ef85,48750ed,38c4b9e,373c5dc,default_qubit_performance -w 12 -d default.qubit time bm_iqp_circuit
```