@johannesjmeyer 2020-04-16
So the new todo list is:
- [ ] Make constants `SQRT2` and `INV_SQRT2` for even less calculations
- [ ] Supersede `diagonal` with `eigvals`
- [ ] Make `_eigvals` and `_matrix` classmethods so that we can redefine `_matrix` in `DiagonalOperation` to use `_eigvals`
- [ ] Unit test `mat_vec_product_einsum`
- [ ] Address the other suggestions