@johannesjmeyer 2020-04-16
By the way, @josh146, `lru_cache` didn't help the performance.