@josh146 2020-04-16
> By the way, @josh146, lru_cache didn't help the performance.

Oh, interesting! I think (not 100% sure) that @trbromley and @antalszava mentioned `lru_cache` provided some speedup for constant matrices, but not for parametrized matrices?