@antalszava 2020-04-16
Sorry for the confusion @josh146! `lru_cache` was considered, but then eventually a further speed-up was achieved with the `@jit` decorator from `numba` (after an initial overhead coming from just-in-time compilation).