@johannesjmeyer 2020-04-30
I didn't end up writing specific tests for `mat_vec_product_einsum`, as the unit tests are all basically testing exactly this. Also `mat_vec_product` didn't have specific tests either. 

I also put `eigvals` at the top of the class hierarchy (into `Operator`) and made both `_eigvals` and `_matrix` classmethods. `_eigvals` now defaults to 
```python
np.linalg.eigvals(cls._matrix(*params))
```

I think we're now pretty close to getting this merged.