@josh146 2020-04-30
Oh, one more thought --- the documentation should also be updated, more specifically the plugin guide here: https://pennylane.readthedocs.io/en/stable/development/plugins.html#supporting-new-operations