@johannesjmeyer 2020-05-02
I also honestly don't know what CodeFactor complains with this error. The method is not marked as abstract.