@josh146 2020-05-05
>I also honestly don't know what CodeFactor complains with this error. The method is not marked as abstract.

Should be fixed now @johannesjmeyer!

@co9olguy and @antalszava, are the responses to the eigenvalues comments agreeable to merging?