@antalszava 2020-05-05
Sure! (no changes requested from my part)

The lines for what exactly is a classmethod/property, which private method depends on which public are blurring pretty much. This will probably cause some headaches in the future. Also, I'd be in for removing `eigvals` definitions which are not necessary internally (but this is subjective).

Having said that I understand and agree with the changes in here (doesn't seem so easy to tidy it further in this PR).