@antalszava 2020-05-06
Looks good to me, the docstrings are great and descriptive! :100: